requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
python-dotenv==0.19.0
numpy==1.21.2
opencv-python==4.5.3.56
pygame==2.0.1
python-multipart==0.0.5
aiofiles==0.7.0
Pillow==8.3.1
pydantic[email]
email-validator
twilio
python-jose==3.3.0
passlib==1.7.4